+++ graphql/src/main/java/com/github/yituhealthcare/arc/graphql/rest/GraphQLController.java
@@ -1,10 +1,10 @@
 package com.github.yituhealthcare.arc.graphql.rest;
 
+import brave.http.HttpResponseParser;
 import com.github.yituhealthcare.arc.graphql.GraphQLProvider;
 import com.github.yituhealthcare.arc.graphql.event.DomainEvent;
 import com.github.yituhealthcare.arc.mq.Message;
 import com.github.yituhealthcare.arc.mq.producer.Producer;
-import brave.http.HttpResponseParser;
 import graphql.ExecutionInput;
 import graphql.ExecutionResult;
 import graphql.GraphQLContext;
@@ -16,9 +16,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @RestController
 public class GraphQLController {
@@ -37,7 +40,7 @@ public class GraphQLController {
     private boolean eventEnable;
 
     @PostMapping(value = "${arc.graphql.path:graphql}", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public Object dispatcher(@RequestBody GraphQLRequestBody requestBody, HttpServletResponse response) {
+    public Object dispatcher(@RequestBody GraphQLRequestBody requestBody, HttpServletRequest request, HttpServletResponse response) {
         log.info("graphql receive body:{}", requestBody);
         if ("IntrospectionQuery".equalsIgnoreCase(requestBody.getOperationName())) {
             // schema 查询直接return
@@ -47,7 +50,12 @@ public class GraphQLController {
                 .query(requestBody.getQuery())
                 .operationName(requestBody.getOperationName())
                 .variables(requestBody.getVariables())
-                .context(GraphQLContext.newContext().of("query", requestBody.getQuery()).build())
+                .context(GraphQLContext.newContext()
+                        .of("query", requestBody.getQuery())
+                        .of("headers", Collections.list(request.getHeaderNames())
+                                .stream()
+                                .collect(Collectors.toMap(h -> h, request::getHeader)))
+                        .build())
                 .build();
 
         ExecutionResult executionResult = graphQLProvider.getGraphQL().execute(executionInput);